refactor: move useSelect to its own folder.
