+++ docs/data/base/components/select/UseSelect.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { useSelect, SelectOption } from '@mui/base';
+import useSelect from '@mui/base/useSelect';
+import { SelectOption } from '@mui/base';
 import { styled } from '@mui/system';
 
 const grey = {
