+++ packages/mui-base/src/SelectUnstyled/SelectUnstyledContext.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { OptionState } from '@mui/base/useListbox';
-import { SelectOption, UseSelectOptionSlotProps } from './useSelect.types';
+import { SelectOption, UseSelectOptionSlotProps } from '@mui/base/useSelect';
 
 export interface SelectUnstyledContextType {
   getOptionState: (value: SelectOption<any>) => OptionState;
