+++ packages/mui-base/src/SelectUnstyled/defaultOptionStringifier.ts
@@ -1,4 +1,4 @@
-import { SelectOption } from './useSelect.types';
+import { SelectOption } from '@mui/base/useSelect';
 
 const defaultOptionStringifier = <TValue>(option: SelectOption<TValue>) => {
   const { label, value } = option;
