+++ packages/mui-base/src/SelectUnstyled/index.ts
@@ -7,8 +7,4 @@ export * from './selectUnstyledClasses';
 
 export * from './SelectUnstyled.types';
 
-export { default as useSelect } from './useSelect';
-
-export * from './useSelect.types';
-
 export * from './utils';
