+++ packages/mui-base/src/useSelect/index.ts
@@ -1,0 +1,5 @@
+export { default as useSelect } from './useSelect';
+export { default } from './useSelect';
+export * from './useSelect';
+
+export * from './useSelect.types';
